Python_app/requirements.txt
mysql-connector-python>=8.0
sv-ttk>=2.6.0
